requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
sentence-transformers
requests